moved EventOverview tests into a folder
